Update decorator-di.md - Round #1
I re-worded A LOT throughout, so please double check me and make sure everything still means the same thing. There was a moment or two that, even with the audio, I couldn't really understand what was being said, so I did the best I could.

I also have some fun ideas for the visuals at the end of this video. Maybe a safari person "spotting patterns in the wild" or something. XD