Update command-real-world.md - Round #1
I didn't have the video to go on for this one, but it was fairly straightforward. May need to double check things like names for accurate capitalization.

We also haven't really used a bulleted list like that in a script before. I don't think that's necessarily a bad thing, but you guys should discuss whether you want to keep it that way, or convert that info to paragraph form. I can totally turn it into a bulleted list graphic for the video though.

Line 5 may need to change a little to flow into the next paragraph a little better.