Update cor-conclusion-and-middleware.md - Round #1 (#29)
* Update cor-conclusion-and-middleware.md - Round #1

I don't think I've handled a script that used check and X emojis, but I do like the way it looks, so I left it in. I just added them at the beginning so it looks more like bullet points, and changed the sentence structure of each item a little so Kevin can read them like they're just part of an ongoing paragraph. I'll probably list these out on the screen in a similar format when we get around to the video edit.

* break paragraphs into multiple lines

---------

Co-authored-by: Diego Aguiar <barret91@hotmail.com>