State Coding (#31)
* add script file

* Update state-coding.md - Round #1 (#34)

* Update state-coding.md - Round #1

This was a longer one and there's quite a bit to cover. I reworded a lot to improve flow. Definitely make sure it's accurate. I also removed that link. I don't remember seeing links in past scripts I've worked on, so if that needs to be included, I'd ask Ryan or decide how you'd like to handle those going forward. My focus going through these is manly on what Kevin will say aloud.

I've also noticed that the script text varies quite a bit from what is said in the original recording, which can complicate things. Especially when a bunch of stuff is happening on the screen (maybe something repetitive) and you verbally walk through it in the recording, but omit that in the script text. If I don't add something back to the script to fill the void, then there's just a bunch of silence while very necessary coding is happening on the screen. I can only speed the video up so much before it's unintelligible (and, consequently, un-useful), so keep that in mind.

* add some tweaks to the script file

---------

Co-authored-by: Diego Aguiar <barret91@hotmail.com>

---------

Co-authored-by: Danielle <berealady9@gmail.com>