Factory Families (#37)
* add factory-families.md script

* Update factory-families.md - Round #1 (#40)

* Update factory-families.md - Round #1

This needs the usual double checking. There's some stuff in line 25 that I left because I wasn't 100% sure about it (the link and the [ ]).

Diego: The misspelling of "Konami" is likely fixable if you can give me a screenshot of your screen with that line spelled correctly. I can cut the line out and add it over the video file.

* some tweaks to the script file

* more tweaks

---------

Co-authored-by: Diego Aguiar <barret91@hotmail.com>

* Update factory-families.md (#41)

* fix optional value for input

---------

Co-authored-by: Danielle <berealady9@gmail.com>
Co-authored-by: Kevin Bond <kevinbond@gmail.com>